tela de cadastrar profissional agora está funcional

+++ src/main/java/com/pi/logic/service/TreinoService.java
@@ -103,9 +103,11 @@ public class TreinoService {
 
         List<TreinoEntity> treinos = treinoRepository.treinosDoCliente(clienteID);
 
-        treinos.forEach((treino) -> {
-            treinoRepository.delete(treino);
-        });
+        if (!treinos.isEmpty()) {
+            treinos.forEach((treino) -> {
+                treinoRepository.delete(treino);
+            });
+        }
 
     }
